api/requirements.txt
flask>=3.0
flask-cors>=4.0
langchain>=0.3
langchain-openai>=0.2
langchain-community>=0.3
langchain-text-splitters>=0.3
faiss-cpu>=1.8
openai>=1.0
python-dotenv>=1.0
tiktoken>=0.7
gunicorn>=22.0
